requirements.txt
torch>=2.2.0
transformers>=4.39.0
datasets>=2.18.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
PyYAML>=6.0.1
flake8>=7.0.0
radon>=6.0.1
flask>=3.0.0
